支持Fragment自动实例和加载 //当已存在复用，不存在则创建

+++ edge/src/main/java/com/shuanglu/edge/Model/Database/Table/DBTDownload.kt
@@ -124,7 +124,7 @@ class DBTDownload {
                     cursor.getString(cursor.getColumnIndex(FILE_LOCAL_PATH)),
                     cursor.getString(cursor.getColumnIndex(FILE_TEMP_PATH))
                 )
-                model.id = cursor.getInt(cursor.getColumnIndex(ID))
+                model.id = cursor.getLong(cursor.getColumnIndex(ID))
                 model.totalSize = cursor.getLong(cursor.getColumnIndex(FILE_TOTAL_SIZE))
                 model.alreadyDownSize = cursor.getLong(cursor.getColumnIndex(FILE_ALREAD_DOWN_SIZE))
                 list.add(edgeDownloadModel)
